Add Task5.java and edit Task3.java

+++ 1semester/tasks_from_study.istamendil.info/Task3.java
@@ -7,8 +7,8 @@ public class Task3 {
 		System.out.printf("Please, enter nuber of multipliers: ");
 		long n = scan.nextLong();
 		double calculatedPI = 2;
-		for (double i = 1; i <= n; i++) {
-			calculatedPI *= 4 * i * i / (2 * i - 1) / (2 * i + 1);
+		for (long i = 1; i <= n; i++) {
+			calculatedPI *= 4 * (double)(i * i) / (2 * (double)(i) - 1) / (2 * (double)(i) + 1);
 		}
 		System.out.printf("calculatedPI = %f\nMath.PI = %f\n", calculatedPI, Math.PI);
 	}
